Replace type in tick.h

+++ lab_07_2/tick.h
@@ -1,6 +1,6 @@
 #ifndef TICK_H
 #define TICK_H
 
-unsigned long int tick(void);
+unsigned long long tick(void);
 
 #endif
